move RebuttalForm componentent into scenes/game
