+++ src/scenes/game/Game.tsx
@@ -17,7 +17,7 @@ import {
 import { EndGame } from "src/scenes/game/EndGame";
 import { Header } from "src/scenes/game/Header";
 import { Hint } from "src/scenes/game/Hint";
-import { RebuttalForm } from "src/scenes/RebuttalForm";
+import { RebuttalForm } from "src/scenes/game/RebuttalForm";
 import {
   Action,
   ActionTypes,
